hw5/hw5/requirements.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.19.5
Pillow==8.1.0
scipy==1.9.1
tqdm==4.64.1
sklearn==1.1.2